Finished country population updater, changed country population to long.

+++ Covid-19 Tracker/Covid-19 Tracker/Model/Country.cs
@@ -7,7 +7,7 @@ namespace Covid_19_Tracker.Model
         public int Id { get; set; }
         public string Name { get; set; }
         public string IsoCode { get; set; }
-        public int Population { get; set; }
+        public long Population { get; set; }
         public List<Infected> Infected { get; set; }
         public List<Vaccinated> Vaccinated { get; set; }
     }
